checksum: structure module
